Removes cruft
