+++ app/components/search/Sidebar/Sidebar.tsx
@@ -1,4 +1,4 @@
-import React, { useState, useMemo } from "react";
+import React, { useState } from "react";
 
 import type { Category } from "models/Meta";
 
@@ -10,7 +10,6 @@ import ClearAllFilters from "components/search/Refinements/ClearAllFilters";
 import OpenNowFilter from "components/search/Refinements/OpenNowFilter";
 import RefinementListFilter from "components/search/Refinements/RefinementListFilter";
 import FacetRefinementList from "components/search/Refinements/FacetRefinementList";
-import { eligibilityMap as ucsfEligibilityMap } from "components/ucsf/RefinementLists/ucsfEligibilitiesMap";
 import { Button } from "components/ui/inline/Button/Button";
 import MobileMapToggleButtons from "./MobileMapToggleButtons";
 import styles from "./Sidebar.module.scss";
@@ -20,7 +19,6 @@ const Sidebar = ({
   searchRadius,
   isSearchResultsPage,
   eligibilities = [],
-  categorySlug = "",
   subcategories = [],
   subcategoryNames = [],
   sortAlgoliaSubcategoryRefinements = false,
@@ -31,7 +29,6 @@ const Sidebar = ({
   searchRadius: string;
   isSearchResultsPage: boolean;
   eligibilities?: object[];
-  categorySlug?: string;
   subcategories?: Category[];
   subcategoryNames?: string[];
   sortAlgoliaSubcategoryRefinements?: boolean;
@@ -67,25 +64,6 @@ const Sidebar = ({
     setSearchRadius(evt.target.value);
   };
 
-  const eligibilityNames = useMemo(() => {
-    // This var holds a set of eligibility names pertaining to a given whitelabeled category. This
-    // set can be used to filter the collection of eligibilities returned by Algolia in cases where
-    // we only want to display a specific list of eligibilities for said whitelabeled category.
-
-    // This only accepts UCSF specific slugs because UCSF is currently the only whitelabel that
-    // has a defined and limited set of eligibility filters. In the future, we can create a
-    // universal map of slugs to eligibility sets across all whitelabels.
-    const resourceEligibilityGroups = ucsfEligibilityMap[categorySlug];
-    if (!resourceEligibilityGroups) {
-      return [];
-    }
-
-    const flatEligibilities = resourceEligibilityGroups.flatMap(
-      (group) => group.eligibilities
-    );
-    return flatEligibilities.map((eligibility) => eligibility.name);
-  }, [categorySlug]);
-
   // Currently, the Search Results Page uses generic categories/eligibilities while the
   // Service Results Page uses COVID-specific categories. This logic determines which
   // of these to use as based on the isSearchResultsPage value
@@ -110,32 +88,10 @@ const Sidebar = ({
       eligibilityRefinementJsx = (
         <RefinementListFilter
           attribute="eligibilities"
-          /*
-            The eligibilityNames array represents a static list of eligibilites that are displayed
-            in the ServiceDiscovery form of the parent tile category. (Currently, only the UCSF
-            whitelabel uses this mechanism.)
-
-            If the eligibilityNames array is > 0, we use it to filter out Algolia's returned set
-            of eligibilities. For such cases, we want Algolia to return a large number of
-            eligiblities to ensure that the returned eligibilities include those on the static list.
-            If eligibilityNames is < 1, we just pass Algolia's default limit, 10, and accept those
-            10 without filtering them.
-
-            N.B.: Eligibilities, and refinements in general, are returned by Algolia in order of
-            `[count:desc`, name:asc`]. In other words, the 10 default eligibilities are the most
-            tagged eligibilities of the returned services, with `name:asc` acting as a tiebreaker.
-          */
           limit={100}
-          transformItems={(items: { label: string }[]) => {
-            let itemsList = items;
-            if (eligibilityNames.length > 0) {
-              itemsList = items.filter(({ label }) =>
-                eligibilityNames.includes(label)
-              );
-            }
-
-            return itemsList.sort(orderByLabel);
-          }}
+          transformItems={(items: { label: string }[]) =>
+            items.sort(orderByLabel)
+          }
         />
       );
     }
