--- app/components/ucsf/ClinicianActions/ClinicianActions.module.scss
@@ -1,41 +1,0 @@
-@import "~styles/utils/_helpers.scss";
-
-.clinicianActions {
-  width: 850px;
-}
-
-.title {
-  font-size: 26px;
-  font-weight: 700;
-  line-height: 1;
-  color: #000;
-}
-
-.modalContent {
-  display: grid;
-  gap: 32px;
-}
-
-.actionContainer {
-  padding: 16px;
-  font-size: 20px;
-  box-shadow: 0px 0px 4px rgba(0, 0, 0, 0.2);
-  border-radius: 4px;
-}
-
-.actionType {
-  font-weight: 700;
-  line-height: 1;
-  padding-bottom: 1em;
-}
-
-.actionList {
-  list-style: initial;
-  list-style-position: inside;
-  padding-left: 10px;
-}
-
-.buttonBar {
-  display: flex;
-  justify-content: end;
-}
