--- app/components/ucsf/Layout/Layout.module.scss
@@ -1,12 +1,0 @@
-@import "app/styles/utils/_helpers.scss";
-
-.grid {
-  display: grid;
-  grid-template-columns: repeat(12, 1fr);
-  grid-column-gap: 16px;
-  max-width: 1450px;
-  padding: 114px 48px 20px;
-  @media screen and (max-width: $break-tablet-s) {
-    padding: 25px 20px 20px;
-  }
-}
