--- app/components/ucsf/Layout/Layout.tsx
@@ -1,11 +1,0 @@
-import React, { ReactNode } from "react";
-
-import styles from "./Layout.module.scss";
-
-export const Layout = ({
-  children,
-  customClass = "",
-}: {
-  children: ReactNode;
-  customClass?: string;
-}) => <div className={`${styles.grid} ${customClass}`}>{children}</div>;
