--- app/components/ucsf/RefinementLists/Refinements.module.scss
@@ -1,43 +1,0 @@
-@import "~styles/utils/_helpers.scss";
-
-.refinementsBox_title {
-  color: $color-grey8;
-  @extend %font-size-large;
-  width: 86px;
-  margin-right: 26px;
-
-  @media screen and (max-width: $break-tablet-s) {
-    width: auto;
-    margin: auto auto 15px;
-  }
-}
-
-.refinementsLabels {
-  display: grid;
-  gap: 45px;
-  list-style: decimal inside;
-}
-
-.refinementsList {
-  display: grid;
-  gap: 8px;
-}
-
-.refinementsGroupLabel {
-  display: list-item;
-}
-
-.refinementsLabel {
-  cursor: pointer;
-  padding-left: 8px;
-}
-
-.listContainer {
-  display: grid;
-  grid-template-columns: 1fr 3fr;
-  gap: 30px;
-
-  @media screen and (max-width: $break-tablet-p) {
-    grid-template-columns: none;
-  }
-}
