--- app/components/ucsf/Section/Section.module.scss
@@ -1,19 +1,0 @@
-@import "app/styles/utils/_helpers.scss";
-
-.title {
-  font-size: 26px;
-  font-weight: 700;
-  color: #000;
-}
-
-.subtitle {
-  font-size: 20px;
-  font-weight: 600;
-  color: #000;
-}
-
-.body {
-  font-size: 16px;
-  color: #404040;
-  margin-top: 8px;
-}
