+++ app/config.ts
@@ -20,7 +20,6 @@ interface Config {
   SENTRY_PUBLIC_KEY: string;
   SFFAMILIES_DOMAIN: string;
   SFFAMILIES_USERWAY_APP_ID: string;
-  UCSF_DOMAIN: string;
   STRAPI_API_TOKEN: string;
   STRAPI_API_URL: string;
 }
@@ -43,7 +42,6 @@ const config: Config = {
   SENTRY_PUBLIC_KEY: "590ba0e5f24b0db2978d0bceb8b45ad3",
   SFFAMILIES_DOMAIN: CONFIG.SFFAMILIES_DOMAIN,
   SFFAMILIES_USERWAY_APP_ID: "4EhNmEIwPz",
-  UCSF_DOMAIN: "dcnav",
   STRAPI_API_TOKEN: CONFIG.STRAPI_API_TOKEN,
   STRAPI_API_URL: CONFIG.STRAPI_API_URL,
 };
