--- app/pages/HomePage/components/Partners/Partners.scss
@@ -1,56 +1,0 @@
-.partners {
-  padding: 50px 0;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  overflow: hidden;
-  margin-top: 70px;
-}
-
-.partnersLists {
-  display: flex;
-  justify-content: center;
-  gap: 20px;
-  width: 90%;
-  max-width: 1600px;
-  @media screen and (max-width: 767px) {
-    flex-direction: column;
-  }
-}
-
-.partners ul {
-  display: flex;
-  flex-wrap: wrap;
-  margin-bottom: 50px;
-  @media screen and (max-width: 767px) {
-    flex-direction: column;
-    align-items: center;
-    justify-items: center;
-  }
-}
-
-.partners li {
-  height: 90px;
-  margin: 15px;
-}
-
-.partners li:not(:first-child) {
-  padding-top: 30px;
-}
-
-.partners li header {
-  font-size: 11px;
-  font-weight: bold;
-  margin-bottom: 15px;
-  color: rgba(0, 0, 0, 0.38);
-}
-
-.partners img {
-  max-height: 100%;
-}
-
-.our415Logo {
-  width: 120px;
-  height: auto;
-}
