+++ config.example.yml
@@ -15,11 +15,6 @@ ALGOLIA_INDEX_PREFIX: "<ENTER_YOUR_GITHUB_USERNAME_HERE>"
 ALGOLIA_APPLICATION_ID: "J8TVT53HPZ"
 ALGOLIA_READ_ONLY_API_KEY: "fdf77b152ff7ce0ea4e4221ff3d17d85"
 
-MOHCD_SUBDOMAIN: "testing"
-# uncomment this to simulate sfserviceguide whitelabel UI locally, assuming that
-# you're pointing to localhost:8080
-# MOHCD_DOMAIN: 'localhost:8080'
-
 # Auth0 config values
 AUTH0_AUDIENCE: "http://localhost:8080/api"
 AUTH0_CLIENT_ID: "UcnuRrX6S0SeDEhW9PRe01wEhcvIRuwc"
